import static dictionary as strings
When this call was changed from `readr::read_csv` to `read.csv` in commit a398f7b3adda566fb13759fa04e17a0cb2df319b we lost the `stringsAsFactors = FALSE`, which is the default in the former and not in the latter. This makes `stringsAsFactors = FALSE` explicit, which I believe is a lot safer in general (and resolves a small problem I was seeing when using this code in a different context).