Add uncompressed codelist and panel
This adds CSVs for codelist and codelist_panel that should be pretty much carbon copies of the RDAs, except for CLDR names. The purpose is to make it easier to see diffs in version control, to verify that we aren't making unintended changes to the RDAs.

This solves the issue mentioned in https://github.com/vincentarelbundock/countrycode/issues/16#issuecomment-413822898.